Lofreq patch

+++ modules/lofreq/1.0/lofreq.smk
@@ -1,12 +1,11 @@
 #!/usr/bin/env snakemake
 
-
 ##### ATTRIBUTION #####
 
 
 # Original Author:  Bruno Grande
 # Module Author:    Bruno Grande
-# Contributors:     N/A
+# Contributors:     Kostia Dreval, Ryan Morin
 
 
 ##### SETUP #####
@@ -22,6 +21,9 @@ CFG = op.setup_module(
     version = "1.0",
     subdirectories = ["inputs", "lofreq", "combined", "filtered", "outputs"],
 )
+#set variable for prepending to PATH based on config
+SCRIPT_PATH = CFG['inputs']['src_dir']
+#this is used in place of the shell.prefix() because that was not working consistently. This is not ideal. 
 
 # Define rules to be run locally when using a compute cluster
 localrules:
@@ -52,7 +54,8 @@ rule _lofreq_run:
         tumour_bam = CFG["dirs"]["inputs"] + "bam/{seq_type}--{genome_build}/{tumour_id}.bam",
         normal_bam = CFG["dirs"]["inputs"] + "bam/{seq_type}--{genome_build}/{normal_id}.bam",
         fasta = reference_files("genomes/{genome_build}/genome_fasta/genome.fa"),
-        dbsnp = reference_files("genomes/{genome_build}/variation/dbsnp.common_all-151.vcf.gz")
+        dbsnp = reference_files("genomes/{genome_build}/variation/dbsnp.common_all-151.vcf.gz"), #in our experience, this filter doesn't remove as many SNPs as one would expect
+        bed = reference_files("genomes/{genome_build}/genome_fasta/main_chromosomes.bed")
     output:
         vcf_snvs_filtered = CFG["dirs"]["lofreq"] + "{seq_type}--{genome_build}/{tumour_id}--{normal_id}--{pair_status}/somatic_final_minus-dbsnp.snvs.vcf.gz",
         vcf_indels_filtered = CFG["dirs"]["lofreq"] + "{seq_type}--{genome_build}/{tumour_id}--{normal_id}--{pair_status}/somatic_final_minus-dbsnp.indels.vcf.gz",
@@ -73,16 +76,22 @@ rule _lofreq_run:
         **CFG["resources"]["lofreq"]
     shell:
         op.as_one_line("""
-        if [ ! -e {output.vcf_snvs_filtered} ] && [ -e {params.rm_files} ]; then rm $(dirname {output.vcf_snvs_all})/*; fi
-          &&
-        lofreq somatic {params.opts} --threads {threads} -t {input.tumour_bam} -n {input.normal_bam}
-        -f {input.fasta} -o $(dirname {output.vcf_snvs_filtered})/ -d {input.dbsnp} {params.regions} 
-        > {log.stdout} 2> {log.stderr}
+        SCRIPT_PATH={SCRIPT_PATH};
+        PATH=$SCRIPT_PATH:$PATH;
+        SCRIPT="$SCRIPT_PATH/lofreq2_call_pparallel.py";
+        if [[ $(which lofreq2_call_pparallel.py) =~ $SCRIPT ]]; then 
+            echo "using bundled patched script $SCRIPT";
+            if [ ! -e {output.vcf_snvs_filtered} ] && [ -e {params.rm_files} ]; then rm $(dirname {output.vcf_snvs_all})/*; fi
             &&
-        rm {params.rm_files}
+            lofreq somatic {params.opts} --threads {threads} -t {input.tumour_bam} -n {input.normal_bam}
+            -f {input.fasta} -o $(dirname {output.vcf_snvs_filtered})/ -d {input.dbsnp} {params.regions} --bed {input.bed}
+            > {log.stdout} 2> {log.stderr}
+            &&
+            rm {params.rm_files};
+        else echo "WARNING: PATH is not set properly, using $(which lofreq2_call_pparallel.py)"; fi
         """)
 
-
+# indels are not yet called but this rule merges the empty indels file with the snvs file to produce the consistently named "combined" vcf. 
 rule _lofreq_combine_vcf:
     input:
         vcf_all = expand(CFG["dirs"]["lofreq"] + "{{seq_type}}--{{genome_build}}/{{tumour_id}}--{{normal_id}}--{{pair_status}}/somatic_final.{var_type}.vcf.gz",
@@ -95,7 +104,7 @@ rule _lofreq_combine_vcf:
     resources:
         **CFG["resources"]["bcftools_sort"]
     conda:
-        CFG["conda_envs"]["bcftools"]
+        CFG["conda_envs"]["lofreq"]
     log:
         stdout_all = CFG["logs"]["combined"] + "{seq_type}--{genome_build}/{tumour_id}--{normal_id}--{pair_status}/lofreq_final.combined.stdout.log",
         stderr_all = CFG["logs"]["combined"] + "{seq_type}--{genome_build}/{tumour_id}--{normal_id}--{pair_status}/lofreq_final.combined.stderr.log",
@@ -116,8 +125,7 @@ rule _lofreq_combine_vcf:
 rule _lofreq_filter_vcf:
     input:
         vcf_all = rules._lofreq_combine_vcf.output.vcf_all,
-        vcf_all_filtered = rules._lofreq_combine_vcf.output.vcf_all_filtered,
-        lofreq_filter = CFG["inputs"]["lofreq_filter"]
+        vcf_all_filtered = rules._lofreq_combine_vcf.output.vcf_all_filtered
     output:
         vcf_all_clean = CFG["dirs"]["filtered"] + "{seq_type}--{genome_build}/{tumour_id}--{normal_id}--{pair_status}/somatic_final.combined.filtered.vcf.gz",
         vcf_all_filtered_clean = CFG["dirs"]["filtered"] + "{seq_type}--{genome_build}/{tumour_id}--{normal_id}--{pair_status}/somatic_final_minus-dbsnp.combined.filtered.vcf.gz",
@@ -126,13 +134,16 @@ rule _lofreq_filter_vcf:
     resources:
         **CFG["resources"]["bcftools_sort"]
     conda:
-        CFG["conda_envs"]["bcftools"]
+        CFG["conda_envs"]["lofreq"]
     shell:
         op.as_one_line("""
-        bash {input.lofreq_filter} {input.vcf_all} | bgzip > {output.vcf_all_clean}
+        PATH={SCRIPT_PATH}:$PATH;
+        SCRIPT=$(which lofreq_filter.sh); 
+        echo "using bundled custom filtering script $SCRIPT";
+        lofreq_filter.sh {input.vcf_all} | bgzip > {output.vcf_all_clean}
           && tabix -p vcf {output.vcf_all_clean}
               &&
-        bash {input.lofreq_filter} {input.vcf_all_filtered} | bgzip > {output.vcf_all_filtered_clean}
+        lofreq_filter.sh {input.vcf_all_filtered} | bgzip > {output.vcf_all_filtered_clean}
           && tabix -p vcf {output.vcf_all_filtered_clean}
         """)
